Moved flutter const to new module
