+++ src/auto_flutter_dig/module/flutter/model/_const.py
